+++ cloud/workflow-server/src/main/kotlin/pers/acp/admin/workflow/base/BaseWorkFlowDomain.kt
@@ -1,0 +1,5 @@
+package pers.acp.admin.workflow.base
+
+import pers.acp.admin.common.base.BaseDomain
+
+abstract class BaseWorkFlowDomain : BaseDomain()