+++ cloud/workflow-server/src/main/kotlin/pers/acp/admin/workflow/base/BaseWorkFlowNotify.kt
@@ -1,0 +1,9 @@
+package pers.acp.admin.workflow.base
+
+import pers.acp.spring.boot.exceptions.ServerException
+import kotlin.jvm.Throws
+
+interface BaseWorkFlowNotify {
+    @Throws(ServerException::class)
+    fun doNotify(taskId: String, userId: String)
+}