+++ cloud/workflow-server/src/main/kotlin/pers/acp/admin/workflow/base/PendingCreatedNotify.kt
@@ -1,0 +1,3 @@
+package pers.acp.admin.workflow.base
+
+interface PendingCreatedNotify : BaseWorkFlowNotify